SE-4: load event details from server, migrate IDs to String
- Add GET /event/orienteering/competitions/public/{id} endpoint (backend)
- Replace mock getEventDetails with real network call
- Migrate CyclicEventDetails.eventId and EventParticipantGroup.groupId to String
- Migrate navigation routes (EventDetailsRoute, EventParticipantGroupRoute, EventResultsRoute) to String
- Fix EventsScreen: pass competition.remoteId instead of list index

+++ feature/events/src/main/java/com/rodionov/events/presentation/main/EventsViewModel.kt
@@ -29,9 +29,7 @@ class EventsViewModel(
             is EventsAction.EventClick -> {
                 viewModelScope.launch {
                     navigation.navigate(
-                        EventsNavigation.EventDetailsRoute(
-                            eventId = action.eventId.toLong()
-                        )
+                        EventsNavigation.EventDetailsRoute(eventId = action.eventId)
                     )
                 }
             }
